Support TBB version >= 2021 and drop support for Python 3.6
